+++ src/civitconfig/__main__.py
@@ -5,7 +5,7 @@ import traceback
 from operator import itemgetter
 
 
-from civitdl.helper.utils import run_in_dev
+from helpers.utils import run_in_dev
 from civitconfig.data.setupJson import setupJson
 from civitconfig.args.argparser import get_args
