+++ src/civitdl/__main__.py
@@ -5,7 +5,7 @@ import traceback
 from operator import itemgetter
 
 
-from .helper.utils import run_in_dev
+from helpers.utils import run_in_dev
 from .batch_download import batch_download
 from civitconfig.data.setupJson import setupJson
 from .args.argparser import get_args
